# frontend/docs-site/site/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Routing Layer \u2014 API Gateway",
    "description": "Unified OpenAI-compatible API gateway for multiple LLM providers.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/admin/overview": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Admin Overview",
        "description": "Return a safe operational snapshot for the Admin Console.",
        "operationId": "admin_overview_v1_admin_overview_get",
        "parameters": [
          {
            "name": "X-Admin-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Admin Overview V1 Admin Overview Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/chat/completions": {
      "post": {
        "tags": [
          "Chat Completions"
        ],
        "summary": "Chat Completions",
        "description": "Handle chat completion requests.\n\nApplies all Phase 1 policy checks before forwarding to the Router:\n1. Rate limiting (enforced by ``enforce_budget`` dependency chain)\n2. Monthly budget cap (enforced by ``enforce_budget``)\n3. Model whitelist (enforced inline below)\n\nThen checks the gateway-level prompt cache (Phase 3 \u2014 Task 3.2)\nbefore forwarding to the Router Engine Service.",
        "operationId": "chat_completions_v1_chat_completions_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatCompletionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatCompletionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/embeddings": {
      "post": {
        "tags": [
          "Embeddings"
        ],
        "summary": "Embeddings",
        "description": "Handle embedding requests.",
        "operationId": "embeddings_v1_embeddings_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmbeddingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbeddingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatChoice": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "message": {
            "$ref": "#/components/schemas/ChatMessage"
          },
          "finish_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finish Reason"
          }
        },
        "type": "object",
        "required": [
          "index",
          "message"
        ],
        "title": "ChatChoice",
        "description": "A single completion choice."
      },
      "ChatCompletionRequest": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "type": "array",
            "title": "Messages"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 2.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature",
            "default": 1.0
          },
          "max_tokens": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Tokens"
          },
          "stream": {
            "type": "boolean",
            "title": "Stream",
            "default": false
          },
          "user": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "provider": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProviderPreferences"
              },
              {
                "type": "null"
              }
            ]
          },
          "response_format": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Format"
          }
        },
        "type": "object",
        "required": [
          "model",
          "messages"
        ],
        "title": "ChatCompletionRequest",
        "description": "OpenAI-compatible chat completion request.\n\nExtended with ``provider`` preferences for intelligent routing."
      },
      "ChatCompletionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "object": {
            "type": "string",
            "title": "Object",
            "default": "chat.completion"
          },
          "created": {
            "type": "integer",
            "title": "Created"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "choices": {
            "items": {
              "$ref": "#/components/schemas/ChatChoice"
            },
            "type": "array",
            "title": "Choices"
          },
          "usage": {
            "$ref": "#/components/schemas/UsageInfo"
          },
          "cache_hit": {
            "type": "boolean",
            "title": "Cache Hit",
            "default": false
          },
          "generation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created",
          "model",
          "provider",
          "choices",
          "usage"
        ],
        "title": "ChatCompletionResponse",
        "description": "Normalised chat completion response (OpenAI-compatible)."
      },
      "ChatMessage": {
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "system",
              "user",
              "assistant",
              "tool"
            ],
            "title": "Role"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              }
            ],
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "ChatMessage",
        "description": "A single message in a chat conversation."
      },
...
      "EmbeddingRequest": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model"
          },
          "input": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              }
            ],
            "title": "Input"
          },
          "user": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User"
          }
        },
        "type": "object",
        "required": [
          "model",
          "input"
        ],
        "title": "EmbeddingRequest",
        "description": "OpenAI-compatible embedding request."
      },
      "EmbeddingResponse": {
        "properties": {
          "object": {
            "type": "string",
            "title": "Object",
            "default": "list"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/EmbeddingVector"
            },
            "type": "array",
            "title": "Data"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "usage": {
            "$ref": "#/components/schemas/UsageInfo"
          }
        },
        "type": "object",
        "required": [
          "data",
          "model",
          "provider",
          "usage"
        ],
        "title": "EmbeddingResponse",
        "description": "Normalised embedding response."
      },
      "EmbeddingVector": {
        "properties": {
          "object": {
            "type": "string",
            "title": "Object",
            "default": "embedding"
          },
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "embedding": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Embedding"
          }
        },
        "type": "object",
        "required": [
          "index",
          "embedding"
        ],
        "title": "EmbeddingVector",
        "description": "A single embedding vector."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ProviderPreferences": {
        "properties": {
          "sort": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "price",
                  "throughput",
                  "latency"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort"
          },
          "order": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order"
          },
          "allow_fallbacks": {
            "type": "boolean",
            "title": "Allow Fallbacks",
            "default": true
          },
          "data_collection": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "allow",
                  "deny"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Collection"
          }
        },
        "type": "object",
        "title": "ProviderPreferences",
        "description": "Client-side routing controls (Phase 3 \u2014 Task 3.1).\n\nAllows clients to dictate the routing strategy on a per-request basis,\nmirroring OpenRouter's ``provider`` object.\n\nAttributes:\n    sort: Optimisation axis \u2014 ``\"price\"``, ``\"throughput\"``, or ``\"latency\"``.\n    order: Ordered list of provider slugs to try (e.g. ``[\"anthropic\", \"openai\"]``).\n    allow_fallbacks: When ``False``, the request fails if the first provider\n        is unavailable rather than falling back to the next in the chain.\n    data_collection: Set to ``\"deny\"`` to exclude providers that may train\n        on user data or do not offer Zero Data Retention (ZDR)."
      },
      "UsageInfo": {
        "properties": {
          "prompt_tokens": {
            "type": "integer",
            "title": "Prompt Tokens",
            "default": 0
          },
          "completion_tokens": {
            "type": "integer",
            "title": "Completion Tokens",
            "default": 0
          },
          "total_tokens": {
            "type": "integer",
            "title": "Total Tokens",
            "default": 0
          },
          "estimated_cost_usd": {
            "type": "number",
            "title": "Estimated Cost Usd",
            "default": 0.0
          },
          "cached_tokens": {
            "type": "integer",
            "title": "Cached Tokens",
            "default": 0
          },
          "cache_discount_usd": {
            "type": "number",
            "title": "Cache Discount Usd",
            "default": 0.0
          },
          "cache_hit": {
            "type": "boolean",
            "title": "Cache Hit",
            "default": false
          }
        },
        "type": "object",
        "title": "UsageInfo",
        "description": "Token usage metadata attached to every completion response."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}